2022/5.auth/python_jwt_auth_example/pyproject.toml
[tool.poetry]
name = "python_jwt_auth_example"
version = "0.1.0"
description = ""
authors = ["yingjie.hu <yingjie.hu@fargowealth.com.hk>"]
packages = [
    { include = "**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyJWT = "^2.3.0"
ipykernel = "^6.9.1"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "main:start"
test = "test:main"